backend/ingestion/requirements.txt
# ==============================================================================
# backend/ingestion/requirements.txt
# Ingestion service dependencies
# ==============================================================================

# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ── Pydantic / Settings ────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.6.1

# ── GCP SDKs ──────────────────────────────────────────────────────────────────
google-cloud-bigquery==3.27.0
google-cloud-storage==2.19.0
google-cloud-documentai==2.31.0
google-cloud-logging==3.11.3
google-cloud-aiplatform==1.74.0
google-cloud-secret-manager==2.21.0
vertexai==1.74.0

# ── Document parsing ──────────────────────────────────────────────────────────
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
xlrd==2.0.1          # Legacy .xls support

# ── Utilities ─────────────────────────────────────────────────────────────────
httpx==0.27.2
